+++ ios/CleverTapReact/CleverTapReactManager.h
@@ -1,10 +1,13 @@
 #import <Foundation/Foundation.h>
+#import "CleverTap.h"
 
 @interface CleverTapReactManager : NSObject
 
 + (instancetype)sharedInstance;
 
 - (void)applicationDidLaunchWithOptions:(NSDictionary *)options;
+
+- (void)setDelegates:(CleverTap *)cleverTapInstance;
 
 @property NSString *launchDeepLink;